[refactor] domain 패키지 클래스 domain.model 패키지로 이동
