+++ src/main/java/junseok/snr/wallet/api/WalletController.java
@@ -4,7 +4,7 @@ package junseok.snr.wallet.api;
 import jakarta.validation.Valid;
 import junseok.snr.wallet.api.dto.CreateWalletRequestDto;
 import junseok.snr.wallet.api.dto.CreateWalletResponseDto;
-import junseok.snr.wallet.domain.Wallet;
+import junseok.snr.wallet.domain.model.Wallet;
 import junseok.snr.wallet.application.service.WalletService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
