+++ src/main/java/junseok/snr/wallet/api/dto/CreateWalletResponseDto.java
@@ -1,6 +1,6 @@
 package junseok.snr.wallet.api.dto;
 
-import junseok.snr.wallet.domain.WalletType;
+import junseok.snr.wallet.domain.model.WalletType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
