+++ src/main/java/junseok/snr/wallet/application/service/EtherWalletService.java
@@ -3,8 +3,8 @@ package junseok.snr.wallet.application.service;
 import io.micrometer.common.util.StringUtils;
 import junseok.snr.wallet.infrastructure.repository.common.Web3jUtils;
 import junseok.snr.wallet.api.dto.CreateWalletRequestDto;
-import junseok.snr.wallet.domain.Wallet;
-import junseok.snr.wallet.domain.WalletType;
+import junseok.snr.wallet.domain.model.Wallet;
+import junseok.snr.wallet.domain.model.WalletType;
 import junseok.snr.wallet.infrastructure.repository.WalletRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
