+++ src/main/java/junseok/snr/wallet/application/service/TransactionService.java
@@ -3,9 +3,9 @@ package junseok.snr.wallet.application.service;
 import jakarta.persistence.criteria.Predicate;
 import junseok.snr.wallet.infrastructure.repository.common.Web3jUtils;
 import junseok.snr.wallet.api.dto.WithdrawDto;
-import junseok.snr.wallet.domain.Transaction;
-import junseok.snr.wallet.domain.TransactionType;
-import junseok.snr.wallet.domain.Wallet;
+import junseok.snr.wallet.domain.model.Transaction;
+import junseok.snr.wallet.domain.model.TransactionType;
+import junseok.snr.wallet.domain.model.Wallet;
 import junseok.snr.wallet.infrastructure.repository.TransactionRepository;
 import junseok.snr.wallet.infrastructure.repository.WalletRepository;
 import lombok.AllArgsConstructor;
