+++ src/main/java/junseok/snr/wallet/application/service/WalletService.java
@@ -1,7 +1,7 @@
 package junseok.snr.wallet.application.service;
 
 import junseok.snr.wallet.api.dto.CreateWalletRequestDto;
-import junseok.snr.wallet.domain.Wallet;
+import junseok.snr.wallet.domain.model.Wallet;
 
 import java.math.BigInteger;
