+++ src/main/java/junseok/snr/wallet/domain/model/TransactionStatus.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.domain;
+package junseok.snr.wallet.domain.model;
 
 public enum TransactionStatus {
     PENDING,
