+++ src/main/java/junseok/snr/wallet/domain/model/Wallet.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.domain;
+package junseok.snr.wallet.domain.model;
 
 
 import jakarta.persistence.*;
