+++ src/main/java/junseok/snr/wallet/domain/model/WalletType.java
@@ -1,0 +1,5 @@
+package junseok.snr.wallet.domain.model;
+
+public enum WalletType {
+    ETH
+}
