+++ src/main/java/junseok/snr/wallet/infrastructure/repository/WalletRepository.java
@@ -1,6 +1,6 @@
 package junseok.snr.wallet.infrastructure.repository;
 
-import junseok.snr.wallet.domain.Wallet;
+import junseok.snr.wallet.domain.model.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WalletRepository extends JpaRepository<Wallet, Long> {
