+++ src/test/java/junseok/snr/wallet/api/domain/TransactionTest.java
@@ -5,10 +5,10 @@ import static org.junit.jupiter.api.Assertions.*;
 import junseok.snr.wallet.application.service.ExceptionCode;
 import junseok.snr.wallet.application.service.TransactionException;
 import junseok.snr.wallet.application.service.WalletException;
-import junseok.snr.wallet.domain.Transaction;
-import junseok.snr.wallet.domain.TransactionStatus;
-import junseok.snr.wallet.domain.TransactionType;
-import junseok.snr.wallet.domain.Wallet;
+import junseok.snr.wallet.domain.model.Transaction;
+import junseok.snr.wallet.domain.model.TransactionStatus;
+import junseok.snr.wallet.domain.model.TransactionType;
+import junseok.snr.wallet.domain.model.Wallet;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
