+++ src/test/java/junseok/snr/wallet/api/domain/WalletTest.java
@@ -1,8 +1,8 @@
 package junseok.snr.wallet.api.domain;
 
 import junseok.snr.wallet.application.service.WalletException;
-import junseok.snr.wallet.domain.Wallet;
-import junseok.snr.wallet.domain.WalletType;
+import junseok.snr.wallet.domain.model.Wallet;
+import junseok.snr.wallet.domain.model.WalletType;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
